requirements.txt
# 核心依赖
pydantic>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Web界面
streamlit>=1.28.0
plotly>=5.15.0

# 向量数据库和RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# 报告生成
reportlab>=4.0.0

# 开发工具
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0

# 部署
docker>=6.0.0
